# server/src/IO.Swagger/Controllers/BlogApi.cs
/*
 * H Plus Sports Company API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class BlogApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="body"></param>
        /// <response code="201">Blog post has been created</response>
        [HttpPost]
        [Route("/blog/create")]
        [ValidateModelState]
        [SwaggerOperation("BlogCreatePost")]
        public virtual IActionResult BlogCreatePost([FromBody]BlogPost body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <response code="200">List of blog articles</response>
        /// <response code="500">Unexpected error</response>
        [HttpGet]
        [Route("/blog")]
        [ValidateModelState]
        [SwaggerOperation("BlogGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Object>), description: "List of blog articles")]
        [SwaggerResponse(statusCode: 500, type: typeof(InlineResponse500), description: "Unexpected error")]
        public virtual IActionResult BlogGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Object>));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InlineResponse500));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Object>>(exampleJson)
                        : default(List<Object>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}